+++ packages/utils/renderer/index.ts
@@ -5,6 +5,7 @@ export * from '../types/intelligence'
 export * from './hooks'
 export * from './ref'
 export * from './storage'
+export * from './window-role'
 
 export * from './touch-sdk'
 export * from './touch-sdk/utils'